Move ContentView to ViewList file
Move all the code from ContentView to a new file called ViewList. That way we can continue with the other development right from Content View
