+++ CWC-Receta-Challenges/CWC-Receta-Challenges/Views/ListViewFormat.swift
@@ -1,0 +1,8 @@
+//
+//  ListViewFormat.swift
+//  CWC-Receta-Challenges
+//
+//  Created by MIGUEL GONZALEZ on 15/1/23.
+//
+
+import Foundation
